CHANGELOG:     AÑADIDOS:         -Añadido mas funcionalidad en el endpoint de publicaciones con filtros         -Simplificando el codigo y diseñando la barra general         -Añadido las funcionalidades a la pagina principal         -Añadido el logo y un placeholder     ELIMINADOS:         ---     NOTAS:         -Pronto añado las tareas de menor prioridad

+++ pages/static/main_page/js/index.js
@@ -1,44 +1,27 @@
 csrf_token = document.getElementsByName('csrfmiddlewaretoken')[0].value
-let button = document.createElement("button");
-button.onclick = function (e) {
+let button = document.getElementById('logout');
+if (button){
+  button.onclick = function (e) {
     fetch("/login", {
       method: "DELETE",
       cache: "no-cache",
       credentials: "same-origin",
       headers: { "X-CSRFToken": csrf_token },
-      body:{
-        csrfmiddlewaretoken: '{{ csrf_token }}'
-      }
     }).then(() => {e.preventDefault(); location.reload()});
   };
-  button.innerHTML = "LOGOUT";
 
-function get_login() { return fetch("/login", {
-  method: "GET",
-  mode: "cors",
-  cache: "no-cache",
-  credentials: "same-origin",
-  headers: {
-    "Content-Type": "application/json",
-  },
-}).then((a) => {
-  if (a.status != 404){
-    return a.json()
-  }
-  console.log("No esat iniciado sesion");
-  
-}).then((data) => data).catch((e) => console.log(e))
 }
 
-get_login().then((data) => {
-  if (data){
-  if (data.nombre) {
-    document.getElementById("nombre").innerHTML = data.nombre;
-    document.getElementById("container").appendChild(button);
+let publicaciones = document.getElementsByClassName('publicacion');
+for (var i = 0; i < publicaciones.length; i++){
+  publicaciones[i].onclick = function(a){
+    location.href = (location.origin + "/publication_page/view/" + this.id);
   }
 }
-});
-    
-  
 
-  
+let buscador = document.getElementById('buscador')
+let buscar = document.getElementById('buscar')
+
+buscar.onclick = function(){
+  location = (location.origin + "/publication_page/filter/1/title/" + buscador.value + "/recent");
+}
